Remove unsupported versions of Rails and add 5.0.7.2
Why is this change needed?
--------------------------
We're updating this gem to support more modern versions of Rails. Rails 5.0.7.2 isn't the target version, but it's the next version beyond what was previously supported.

How does it address the issue?
------------------------------
The tests don't work from this commit alone. The dummy application needs to be updated to be Rails 5/6 compatible, but this removes older versions of Rails so we're not trying to test against long-deprecated software.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------