Determine which version Appraisal is testing before including the dummy application
Why is this change needed?
--------------------------
On the path to improving our testing, we may want to include a different Rails application depending on what version of Rails we're trying to test. To do that, we'll need to know what gemfile Appraisal loaded.

How does it address the issue?
------------------------------
The ENV contains the name of the gemfile. By convention in this project, that contains the Rails version under test. I'm parsing it out in the worst way possible. I expect there to be another pass at this before this code rests.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------