Update dummy application to be Rails 5/6 compatible
Why is this change needed?
--------------------------
We want ActiveMailer to be Rails 6 compatible, so the dummy app for testing needs to support those versions.

How does it address the issue?
------------------------------
I include some updates from Thoughtbot's version in Clearance while maintaining the specific changes needed for this repo. It does not work as-is because Rails 5 expects to find a manifest at app/assets/config/manifest.js and it's not finding it on load even after I create it. Some more debugging will need to happen.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------