Define Rails.root in our dummy application
Why is this change needed?
--------------------------
As the previous commit[1] mentions, the tests weren't running because
of complaints about not finding the manifest file. Sprockets was
checking Rails.root for the existence of the manifest file, like so
```
raise ManifestNeededError unless ::Rails.root.join("app/assets/config/manifest.js").exist?
```
It didn't find the manifest because the dummy application did not have
the Rails root set, so it was looking in the wrong place
(e.g. ~/active_mailer instead of ~/active_mailer/test/fixtures/dummy)

How does it address the issue?
------------------------------
Explicitly set the root in the dummy application

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
1 - 125a836e8f2e6f4698cbe39142af59c58a2d91ad