Remove 5.0.7.2 in favor of 5.1.7
Why is this change needed?
--------------------------
I'm in the process of upgrading this to be compatible and tested with the newest versions of Rails. I'm stepping through the intermediate versions of Rails. When I finish a version, I plan to remove it from support if it is not actively maintained.

How does it address the issue?
------------------------------
- Remove 5.0.7.2 from the Appraisals file
- Remove the 5.0.7.2 Gemfile
- Remove comment from the testing application that claimed to support Rails 3 and 4

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------