Add a manifest file for the testing application
Why is this change needed?
--------------------------
A few commits ago[1] I corrected the testing application so it looked for the manifest in the correct place, but I did not add a default manifest file.

How does it address the issue?
------------------------------
I added a default manifest. The tests now pass.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
1 - 64ddb5f39cbc5685ac41d6e3c896a14409a4b258