Add Gemfile.lock to repo
Why is this change needed?
--------------------------
In the Before Time, when the land was dark, dinosaurs roamed the earth, and I first made this gem, committing the Gemfile.lock was not recommended. Now it's best practice as a way to make sure anyone checking out the repo uses the same gem versions.

How does it address the issue?
------------------------------
Modify the gitignore and add Gemfile.lock

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------